Update CPU and RAM usage history even when the stats window is not shown

+++ src/statswindow.hh
@@ -26,7 +26,7 @@ private:
 
     int coreWinsWidth;
     
-    list<int> cpuUsageHistory, ramUsageHistory;
+    list<double> cpuUsageHistory, ramUsageHistory;
     
     CpuInfo *cpuInfo;
     RamInfo *ramInfo;
